Primeiro refatorando class
